+++ ceammc/ext/src/string/mod_string.cpp
@@ -1,5 +1,6 @@
 #include "mod_string.h"
 #include "string_format.h"
+#include "string_join.h"
 #include "string_str.h"
 #include "string_to_symbol.h"
 
@@ -8,4 +9,5 @@ void ceammc_string_setup()
     string_setup();
     setup_string0x2eto_symbol();
     setup_string0x2eformat();
+    setup_string0x2ejoin();
 }